backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
passlib==1.7.4
python-dotenv==1.0.1
psycopg[binary]==3.2.10
cloudinary==1.36.0
bcrypt==4.1.2
gunicorn==21.2.0



